@@ -3,11 +3,12 @@ apply plugin: "de.mannodermaus.android-junit5"
 apply plugin: "com.hiya.jacoco-android"
 
 // Third party library versions
-def secure_sms_proxy_api_version = '2.1.1'
+def secure_sms_proxy_api_version = '2.2.0'
 def android_billingclient_version = '5.0.0'
 def slidetoact_version = '0.9.0'
 def takisoft_preferencex_version = '1.1.0'
 def junit_version = '5.8.2'
+def work_version = "2.7.1"
 
 jacoco {
     toolVersion = "0.8.4"
@@ -45,11 +46,11 @@ tasks.withType(Test) {
 }
 
 android {
-    compileSdkVersion 31
+    compileSdkVersion 32
     defaultConfig {
         applicationId "com.github.frimtec.android.pikettassist"
         minSdkVersion 26
-        targetSdkVersion 31
+        targetSdkVersion 32
         versionCode vcode
         versionName vname
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -74,6 +75,7 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar', '*.aar'], dir: 'lib')
     implementation 'androidx.preference:preference:1.2.0'
+    implementation "androidx.work:work-runtime:$work_version"
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'com.google.android.material:material:1.6.0'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
