@@ -1,5 +1,5 @@
 package com.github.frimtec.android.pikettassist.donation;
 
 interface RowDataProvider {
-  SkuRowData getData(int position);
+  ProductRowData getData(int position);
 }
