@@ -2,7 +2,6 @@ package com.github.frimtec.android.pikettassist.ui;
 
 import static com.android.billingclient.api.BillingClient.BillingResponseCode;
 import static com.github.frimtec.android.pikettassist.donation.billing.BillingProvider.BillingState.PURCHASED;
-import static com.github.frimtec.android.pikettassist.service.system.Feature.RequestCodes.FROM_BATTERY_OPTIMIZATION_REQUEST_CODE;
 import static com.github.frimtec.android.pikettassist.ui.BillingAdapter.BILLING_DIALOG_TAG;
 import static com.github.frimtec.android.pikettassist.ui.overview.StateFragment.REGISTER_SMS_ADAPTER_REQUEST_CODE;
 
@@ -32,12 +31,11 @@ import androidx.viewpager2.widget.ViewPager2;
 import com.android.billingclient.api.BillingClient;
 import com.github.frimtec.android.pikettassist.R;
 import com.github.frimtec.android.pikettassist.action.Action;
-import com.github.frimtec.android.pikettassist.domain.OnOffState;
 import com.github.frimtec.android.pikettassist.donation.DonationFragment;
 import com.github.frimtec.android.pikettassist.donation.billing.BillingManager;
 import com.github.frimtec.android.pikettassist.donation.billing.BillingProvider;
-import com.github.frimtec.android.pikettassist.service.LowSignalService;
-import com.github.frimtec.android.pikettassist.service.PikettService;
+import com.github.frimtec.android.pikettassist.service.LowSignalWorker;
+import com.github.frimtec.android.pikettassist.service.PikettWorker;
 import com.github.frimtec.android.pikettassist.service.ShiftService;
 import com.github.frimtec.android.pikettassist.service.SmsListener;
 import com.github.frimtec.android.pikettassist.service.system.NotificationService;
@@ -168,7 +166,7 @@ public class MainActivity extends AppCompatActivity {
       registerOnSmsAdapter();
     }
     loadFragment(savedFragmentPosition);
-    PikettService.enqueueWork(this);
+    PikettWorker.enqueueWork(this);
   }
 
   private void registerOnSmsAdapter() {
@@ -216,7 +214,7 @@ public class MainActivity extends AppCompatActivity {
   public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
     super.onRequestPermissionsResult(requestCode, permissions, grantResults);
     refresh();
-    PikettService.enqueueWork(this);
+    PikettWorker.enqueueWork(this);
   }
 
   @Override
@@ -237,8 +235,6 @@ public class MainActivity extends AppCompatActivity {
             registrationErrors[result.getReturnCode().ordinal()];
         Toast.makeText(this, registrationText, Toast.LENGTH_LONG).show();
       }
-    } else if (requestCode == FROM_BATTERY_OPTIMIZATION_REQUEST_CODE) {
-      Log.i(TAG, "Return from grant battery optimization deactivation activity; result=" + resultCode);
     }
   }
 
@@ -331,7 +327,7 @@ public class MainActivity extends AppCompatActivity {
         public void onReceive(Context context, Intent intent) {
           if (Intent.ACTION_AIRPLANE_MODE_CHANGED.equals(intent.getAction()) &&
               new ShiftService(context).getShiftState().isOn()) {
-            LowSignalService.enqueueWork(context);
+            LowSignalWorker.enqueueWork(context);
           }
           refresh();
         }
