@@ -1,8 +1,15 @@
 package com.github.frimtec.android.pikettassist.service;
 
-import android.content.Intent;
+import static com.github.frimtec.android.pikettassist.service.BogusAlarmWorker.ALARM_TYPE_PARAMETER_KEY;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 
-import com.github.frimtec.android.pikettassist.service.BogusAlarmService.AlarmType;
+import androidx.work.Data;
+
+import com.github.frimtec.android.pikettassist.service.BogusAlarmWorker.AlarmType;
 import com.github.frimtec.android.pikettassist.service.system.AlarmService.ScheduleInfo;
 
 import org.junit.jupiter.params.ParameterizedTest;
@@ -13,28 +20,21 @@ import java.util.EnumSet;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.github.frimtec.android.pikettassist.service.BogusAlarmService.INTENT_EXTRA_ALARM_TYPE;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
-import static org.mockito.Mockito.when;
-
-class BogusAlarmServiceWorkUnitTest {
+class BogusAlarmWorkerWorkUnitTest {
 
   @ParameterizedTest
   @EnumSource(AlarmType.class)
   void apply(AlarmType alarmType) {
     // arrange
-    Intent intent = mock(Intent.class);
-    when(intent.getStringExtra(INTENT_EXTRA_ALARM_TYPE)).thenReturn(alarmType.name());
+    Data inputData = new Data.Builder()
+        .putString(ALARM_TYPE_PARAMETER_KEY, alarmType.name())
+        .build();
     Map<AlarmType, Runnable> alarmTriggers = alarmTriggers();
 
-    ServiceWorkUnit workUnit = new BogusAlarmServiceWorkUnit(alarmTriggers);
+    WorkUnit workUnit = new BogusAlarmWorkUnit(alarmTriggers);
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
 
     // assert
     assertThat(scheduleInfo).isEqualTo(Optional.empty());
