@@ -2,6 +2,7 @@ package com.github.frimtec.android.pikettassist.service;
 
 import static android.media.AudioManager.MODE_IN_CALL;
 import static android.media.AudioManager.MODE_NORMAL;
+import static com.github.frimtec.android.pikettassist.service.LowSignalWorkUnit.EXTRA_FILTER_STATE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -15,6 +16,7 @@ import android.content.Intent;
 import android.media.AudioManager;
 
 import androidx.core.util.Pair;
+import androidx.work.Data;
 
 import com.github.frimtec.android.pikettassist.domain.AlertState;
 import com.github.frimtec.android.pikettassist.domain.OnOffState;
@@ -33,7 +35,7 @@ import java.time.Duration;
 import java.time.LocalTime;
 import java.util.Optional;
 
-class LowSignalServiceWorkUnitTest {
+class LowSignalWorkerWorkUnitTest {
 
   @Test
   void applyForPikettStateOffReturnsEmpty() {
@@ -65,7 +67,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -76,11 +78,12 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
 
     // assert
     assertThat(scheduleInfo).isEqualTo(Optional.empty());
@@ -120,7 +123,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -131,18 +134,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -180,7 +184,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -191,18 +195,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofMinutes(15));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -240,7 +245,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -251,18 +256,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -298,7 +304,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -309,18 +315,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService).notifySignalLow(SignalLevel.POOR);
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger).run();
@@ -356,7 +363,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -367,18 +374,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(15);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 15)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(15));
-    verify(resultIntent).putExtra("FILTER_STATE", 30);
+    verify(resultIntent).putExtra(EXTRA_FILTER_STATE, 30);
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -414,7 +422,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -425,18 +433,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(15);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 15)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -473,7 +482,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -484,18 +493,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(15);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 15)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(15));
-    verify(resultIntent).putExtra("FILTER_STATE", 15);
+    verify(resultIntent).putExtra(EXTRA_FILTER_STATE, 15);
     verify(notificationService).notifySignalLow(SignalLevel.OFF);
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger).run();
@@ -531,7 +541,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -542,18 +552,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(45);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 45)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent).putExtra("FILTER_STATE", 46);
+    verify(resultIntent).putExtra(EXTRA_FILTER_STATE, 46);
     verify(notificationService).notifySignalLow(SignalLevel.POOR);
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger).run();
@@ -589,7 +600,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -600,18 +611,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -647,7 +659,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -658,18 +670,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger, never()).run();
@@ -705,7 +718,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -716,18 +729,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService, never()).setVolume(any(Integer.class));
     verify(alarmTrigger).run();
@@ -765,7 +779,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -776,18 +790,19 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
     ScheduleInfo info = scheduleInfo.orElse(new ScheduleInfo(Duration.ZERO));
     Intent resultIntent = mock(Intent.class);
     info.getIntentExtrasSetter().accept(resultIntent);
 
     // assert
     assertThat(info.getScheduleDelay()).isEqualTo(Duration.ofSeconds(90));
-    verify(resultIntent, never()).putExtra(eq("FILTER_STATE"), any(Integer.class));
+    verify(resultIntent, never()).putExtra(eq(EXTRA_FILTER_STATE), any(Integer.class));
     verify(notificationService, never()).notifySignalLow(any(SignalLevel.class));
     verify(volumeService).setVolume(6);
     verify(alarmTrigger, never()).run();
@@ -825,7 +840,7 @@ class LowSignalServiceWorkUnitTest {
     VolumeService volumeService = mock(VolumeService.class);
     Runnable alarmTrigger = mock(Runnable.class);
 
-    ServiceWorkUnit workUnit = new LowSignalServiceWorkUnit(
+    WorkUnit workUnit = new LowSignalWorkUnit(
         applicationPreferences,
         audioManager,
         alertDao,
@@ -836,11 +851,12 @@ class LowSignalServiceWorkUnitTest {
         alarmTrigger,
         context
     );
-    Intent intent = mock(Intent.class);
-    when(intent.getIntExtra("FILTER_STATE", 0)).thenReturn(0);
+    Data inputData = new Data.Builder()
+        .putInt(EXTRA_FILTER_STATE, 0)
+        .build();
 
     // act
-    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(intent);
+    Optional<ScheduleInfo> scheduleInfo = workUnit.apply(inputData);
 
     // assert
     assertThat(scheduleInfo).isEqualTo(Optional.empty());
